Change (#118): Moved hardware-related functions
Moved hardware-related functions from Hanami to its
own library.
Also removed some remaining opencl-connections.
