+++ src/libraries/libKitsunemimiHanamiHardware/README.md
@@ -1,0 +1,5 @@
+# libKitsunemimiHanamiCommon
+
+## Description
+
+Common dataset for the hanami-layer.
