+++ src/libraries/libKitsunemimiHanamiHardware/defaults.pri
@@ -1,0 +1,2 @@
+INCLUDEPATH += $$PWD/src \
+               $$PWD/include
