+++ src/libraries/libKitsunemimiHanamiHardware/libKitsunemimiHanamiHardware.pro
@@ -1,0 +1,11 @@
+TEMPLATE = subdirs
+CONFIG += ordered
+
+SUBDIRS = src
+
+run_tests {
+    SUBDIRS += tests
+
+    tests.depends = src
+}
+
