+++ src/libraries/libKitsunemimiHanamiHardware/tests/tests.pro
@@ -1,0 +1,9 @@
+TEMPLATE = subdirs
+CONFIG += ordered
+QT -= qt core gui
+CONFIG += c++17
+
+SUBDIRS = \
+    unit_tests
+
+tests.depends = src
