+++ src/libraries/libKitsunemimiHanamiHardware/tests/unit_tests/main.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+
+
+int main()
+{
+
+    return 0;
+}
